Moved rest endpoint creation code to separate file
